+++ release.sh
@@ -60,14 +60,14 @@ rm -rf metastore_db/ spark-warehouse/
 ./set-version.sh 3.1.3 2.12.10; mvn clean deploy -Dsign; ./build-whl.sh; ./test-release.sh
 ./set-version.sh 3.2.4 2.12.15; mvn clean deploy -Dsign; ./build-whl.sh; ./test-release.sh
 ./set-version.sh 3.3.4 2.12.15; mvn clean deploy -Dsign; ./build-whl.sh; ./test-release.sh
-./set-version.sh 3.4.3 2.12.17; mvn clean deploy -Dsign; ./build-whl.sh; ./test-release.sh
-./set-version.sh 3.5.1 2.12.18; mvn clean deploy -Dsign; ./build-whl.sh; ./test-release.sh
+./set-version.sh 3.4.4 2.12.17; mvn clean deploy -Dsign; ./build-whl.sh; ./test-release.sh
+./set-version.sh 3.5.3 2.12.18; mvn clean deploy -Dsign; ./build-whl.sh; ./test-release.sh
 rm -rf python/dist
 
 ./set-version.sh 3.2.4 2.13.5; mvn clean deploy -Dsign; ./test-release.sh
 ./set-version.sh 3.3.4 2.13.8; mvn clean deploy -Dsign; ./test-release.sh
-./set-version.sh 3.4.3 2.13.8; mvn clean deploy -Dsign; ./test-release.sh
-./set-version.sh 3.5.1 2.13.8; mvn clean deploy -Dsign; ./test-release.sh
+./set-version.sh 3.4.4 2.13.8; mvn clean deploy -Dsign; ./test-release.sh
+./set-version.sh 3.5.3 2.13.8; mvn clean deploy -Dsign; ./test-release.sh
 rm -rf metastore_db/ spark-warehouse/
 
 # all SNAPSHOT versions build, test and complete the example, releasing
@@ -105,17 +105,17 @@ echo
 # create release
 echo "Creating release packages"
 mkdir -p python/pyspark/jars/
-./set-version.sh 3.0.3 2.12.10; mvn clean deploy -Dsign; ./build-whl.sh
-./set-version.sh 3.1.3 2.12.10; mvn clean deploy -Dsign; ./build-whl.sh
-./set-version.sh 3.2.4 2.12.15; mvn clean deploy -Dsign; ./build-whl.sh
-./set-version.sh 3.3.4 2.12.15; mvn clean deploy -Dsign; ./build-whl.sh
-./set-version.sh 3.4.2 2.12.17; mvn clean deploy -Dsign; ./build-whl.sh
-./set-version.sh 3.5.0 2.12.18; mvn clean deploy -Dsign; ./build-whl.sh
+./set-version.sh 3.0.3 2.12.10; mvn clean deploy -Dsign -Dspotless.check.skip -DskipTests -Dmaven.test.skip=true; ./build-whl.sh
+./set-version.sh 3.1.3 2.12.10; mvn clean deploy -Dsign -Dspotless.check.skip -DskipTests -Dmaven.test.skip=true; ./build-whl.sh
+./set-version.sh 3.2.4 2.12.15; mvn clean deploy -Dsign -Dspotless.check.skip -DskipTests -Dmaven.test.skip=true; ./build-whl.sh
+./set-version.sh 3.3.4 2.12.15; mvn clean deploy -Dsign -Dspotless.check.skip -DskipTests -Dmaven.test.skip=true; ./build-whl.sh
+./set-version.sh 3.4.4 2.12.17; mvn clean deploy -Dsign -Dspotless.check.skip -DskipTests -Dmaven.test.skip=true; ./build-whl.sh
+./set-version.sh 3.5.3 2.12.18; mvn clean deploy -Dsign -Dspotless.check.skip -DskipTests -Dmaven.test.skip=true; ./build-whl.sh
 
-./set-version.sh 3.2.4 2.13.5; mvn clean deploy -Dsign
-./set-version.sh 3.3.4 2.13.8; mvn clean deploy -Dsign
-./set-version.sh 3.4.2 2.13.8; mvn clean deploy -Dsign
-./set-version.sh 3.5.0 2.13.8; mvn clean deploy -Dsign
+./set-version.sh 3.2.4 2.13.5; mvn clean deploy -Dsign -Dspotless.check.skip -DskipTests -Dmaven.test.skip=true
+./set-version.sh 3.3.4 2.13.8; mvn clean deploy -Dsign -Dspotless.check.skip -DskipTests -Dmaven.test.skip=true
+./set-version.sh 3.4.4 2.13.8; mvn clean deploy -Dsign -Dspotless.check.skip -DskipTests -Dmaven.test.skip=true
+./set-version.sh 3.5.3 2.13.8; mvn clean deploy -Dsign -Dspotless.check.skip -DskipTests -Dmaven.test.skip=true
 
 # upload to test PyPi
 pip install twine